pyproject.toml
[tool.poetry]
name = "reloc"
version = "1.3.0"
description = "Private owned URL shortener"
authors = ["Andrew Barrientos <andrew@ddotb.tk>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.0"
flask-simplelogin = "^0.0.7"
pytz = "^2021.1"
WTForms = "^2.3.3"
flask-wtf = "^0.15.1"
Flask-Bootstrap4 = "^4.0.2"
Flask-Compress = "^1.10.1"
email-validator = "^1.1.3"
python-gnupg = "^0.4.7"
pymongo = "^3.12.0"
pycryptodome = "^3.10.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
